+++ src/audio-instruments/audio-cg/index.ts
@@ -13,11 +13,10 @@ import E4_mp3 from "./E4.mp3";
 import A4_mp3 from "./A4.mp3";
 import E5_mp3 from "./E5.mp3";
 import A5_mp3 from "./A5.mp3";
-import { GenericInstrument } from "../generic-instrument";
 
 // Use direct path to instrument.ts in audio module.
 // Instrument modules must not depend on the audio module.
-import { Instrument } from "../../audio/instrument";
+import { InstrumentSamples } from "../../audio/instrument";
 
 /**
  * Export classical guitar instrument object.
@@ -30,16 +29,26 @@ import { Instrument } from "../../audio/instrument";
  *   Audio.addInstrument(ClassicalGuitar);
  * ```
  */
-const ClassicalGuitar: Instrument = new GenericInstrument("Classical Guitar", {
-    "E2": E2_mp3,
-    "A2": A2_mp3,
-    "D3": D3_mp3,
-    "G3": G3_mp3,
-    "B3": B3_mp3,
-    "E4": E4_mp3,
-    "A4": A4_mp3,
-    "E5": E5_mp3,
-    "A5": A5_mp3
-});
+class ClassicalGuitarSamples implements InstrumentSamples {
+    getName(): string {
+        return "Classical Guitar";
+    }
+
+    getSamples(): Record<string, string> {
+        return {
+            "E2": E2_mp3,
+            "A2": A2_mp3,
+            "D3": D3_mp3,
+            "G3": G3_mp3,
+            "B3": B3_mp3,
+            "E4": E4_mp3,
+            "A4": A4_mp3,
+            "E5": E5_mp3,
+            "A5": A5_mp3
+        }
+    }
+}
+
+const ClassicalGuitar = new ClassicalGuitarSamples;
 
 export { ClassicalGuitar }