+++ src/audio/instrument.ts
@@ -39,6 +39,23 @@ export interface Instrument {
 }
 
 /**
+ * Instruments can also be added using this interface that holds instrument name and audio samples.
+ */
+export interface InstrumentSamples {
+    /**
+     * Get instrument name.
+     * @returns - Instrument name.
+     */
+    getName(): string;
+
+    /**
+     * Get samples e.g. { "E2": "base64 string mp3", "E3": "base64 string mp3", etc. }
+     * @returns - Record of audio samples.
+     */
+    getSamples(): Record<string, string>;
+}
+
+/**
  * Linear volume to decibels converter.
  * @param linearVolume - Linear volume in range [0, 1].
  * @returns - Volume in decibels.