+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [6.4.0-pre.8] - Unreleased
 ### Added
 - Implemented octave shift annotation kinds "8va", "8vb" and "loco".
+- InstrumentSamples class. Use this for sampler instruments, Classical Guitar for now.
 
 ### Changed
 - Updated annotation layout with lanes, and stuff.